// generated/src/app/api-models/stage-prem-pay-cash-support.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StagePremPayCashSupport {
  batchId: string;
  transactionId: number;
  cashId: number;
  arCashSeqControlBatchId: number;
  arCashControlItemCount: number;
  arCashControlBatchStatus: string;
  arCashControlSecurityCode: string;
  arCashReceiptBatchId: number;
  arCashReceiptCompanyCode: string;
  arCashReceiptGlRefCode: string;
  arCashReceiptSubscriberId: string;
  arCashReceiptDescription: string;
  arCashReceiptStatmntStatus: string;
  arCashReceiptInvoiceNumber: number;
  seqGroupId: number;
  groupType: string;
  groupId: string;
  debitGlNumber: string;
  creditGlNumber: string;
}